+++ builder/common_builder/common_builder.go
@@ -19,7 +19,6 @@ import (
 	"runtime"
 	"strings"
 
-	"github.com/ercole-io/ercole-agent/v2/agentmodel"
 	"github.com/ercole-io/ercole-agent/v2/config"
 	"github.com/ercole-io/ercole-agent/v2/fetcher"
 	"github.com/ercole-io/ercole-agent/v2/logger"
@@ -295,20 +294,4 @@ func CoreFactor(v model.HostData) float64 {
 	}
 
 	return 0.5
-}
-
-func (b *CommonBuilder) RemoveDuplicatedOratabEntries(oratabEntries []agentmodel.OratabEntry) []agentmodel.OratabEntry {
-	m := map[agentmodel.OratabEntry]struct{}{}
-	uniqueOratabEntries := []agentmodel.OratabEntry{}
-
-	for _, d := range oratabEntries {
-		if _, ok := m[d]; !ok {
-			uniqueOratabEntries = append(uniqueOratabEntries, d)
-			m[d] = struct{}{}
-		} else {
-			b.log.Warnf("Duplicated oratab entries %s", d.DBName)
-		}
-	}
-
-	return uniqueOratabEntries
 }
